Allow the context manager to choose a directory
If the context manager in watchdir._watch raises a
watchdir.DestinationDirectoryException, the exception's message is
taken as the destination directory for files at the end of processing.
The default directory is DONE.

WindowMatchingEventHandler currently uses this to populate FAILED when
an Eyes test fails. This is exactly the same as it was before, but it
makes extensions easier.